Rename sideBar to SideBar
