+++ packages/web-app-admin-settings/src/components/Groups/SideBar/EditPanel.vue
@@ -25,7 +25,7 @@
 <script lang="ts">
 import { defineComponent, PropType, ref } from 'vue'
 import { Group } from 'web-client/src/generated'
-import CompareSaveDialog from 'web-pkg/src/components/sideBar/CompareSaveDialog.vue'
+import CompareSaveDialog from 'web-pkg/src/components/SideBar/CompareSaveDialog.vue'
 import { MaybeRef, useClientService } from 'web-pkg'
 import GroupInfoBox from './GroupInfoBox.vue'
