+++ packages/web-app-admin-settings/src/views/Spaces.vue
@@ -88,12 +88,12 @@ import AppTemplate from '../components/AppTemplate.vue'
 import { buildSpace } from 'web-client/src/helpers'
 import { configurationManager } from 'web-pkg'
 import SpacesList from '../components/Spaces/SpacesList.vue'
-import SpaceDetails from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetails.vue'
-import SpaceDetailsMultiple from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetailsMultiple.vue'
-import SpaceNoSelection from 'web-pkg/src/components/sideBar/Spaces/SpaceNoSelection.vue'
+import SpaceDetails from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetails.vue'
+import SpaceDetailsMultiple from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetailsMultiple.vue'
+import SpaceNoSelection from 'web-pkg/src/components/SideBar/Spaces/SpaceNoSelection.vue'
 import ContextActions from '../components/Spaces/ContextActions.vue'
 import MembersPanel from '../components/Spaces/SideBar/MembersPanel.vue'
-import SpaceInfo from 'web-pkg/src/components/sideBar/Spaces/SpaceInfo.vue'
+import SpaceInfo from 'web-pkg/src/components/SideBar/Spaces/SpaceInfo.vue'
 import ActionsPanel from '../components/Spaces/SideBar/ActionsPanel.vue'
 import QuotaModal from 'web-pkg/src/components/Spaces/QuotaModal.vue'
 import { useSideBar } from 'web-pkg/src/composables/sideBar'
