+++ packages/web-app-files/src/fileSideBars.ts
@@ -6,8 +6,8 @@ import SharesPanel from './components/SideBar/Shares/SharesPanel.vue'
 import TagsPanel from './components/SideBar/TagsPanel.vue'
 import NoSelection from './components/SideBar/NoSelection.vue'
 import SpaceActions from './components/SideBar/Actions/SpaceActions.vue'
-import SpaceDetails from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetails.vue'
-import SpaceNoSelection from 'web-pkg/src/components/sideBar/Spaces/SpaceNoSelection.vue'
+import SpaceDetails from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetails.vue'
+import SpaceNoSelection from 'web-pkg/src/components/SideBar/Spaces/SpaceNoSelection.vue'
 import {
   isLocationTrashActive,
   isLocationPublicActive,
@@ -16,11 +16,11 @@ import {
 } from './router'
 import { spaceRoleEditor, spaceRoleManager } from 'web-client/src/helpers/share'
 
-import { Panel } from '../../web-pkg/src/components/sideBar'
+import { Panel } from '../../web-pkg/src/components/SideBar'
 
 import { Resource, User } from 'web-client'
 import { Router } from 'vue-router'
-import SpaceDetailsMultiple from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetailsMultiple.vue'
+import SpaceDetailsMultiple from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetailsMultiple.vue'
 
 function $gettext(msg: string): string {
   return msg
