+++ packages/web-pkg/src/components/SideBar/SideBar.vue
