+++ packages/web-pkg/src/components/SideBar/Spaces/Details/SpaceDetails.vue
