+++ packages/web-pkg/src/components/SideBar/Spaces/Details/SpaceDetailsMultiple.vue
