+++ packages/web-pkg/src/components/SideBar/Spaces/Details/index.ts
