+++ packages/web-pkg/src/components/SideBar/Spaces/SpaceInfo.vue
