+++ packages/web-pkg/tests/unit/components/sidebar/Spaces/Details/SpaceDetails.spec.ts
@@ -1,4 +1,4 @@
-import SpaceDetails from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetails.vue'
+import SpaceDetails from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetails.vue'
 import { spaceRoleManager, ShareTypes } from 'web-client/src/helpers/share'
 import {
   createStore,
