+++ packages/web-pkg/tests/unit/components/sidebar/Spaces/Details/SpaceDetailsMultiple.spec.ts
@@ -1,4 +1,4 @@
-import SpaceDetailsMultiple from 'web-pkg/src/components/sideBar/Spaces/Details/SpaceDetailsMultiple.vue'
+import SpaceDetailsMultiple from 'web-pkg/src/components/SideBar/Spaces/Details/SpaceDetailsMultiple.vue'
 import { defaultPlugins, shallowMount } from 'web-test-helpers'
 
 const spaceMock = {
