+++ packages/web-pkg/tests/unit/components/sidebar/Spaces/SpaceInfo.spec.ts
@@ -1,4 +1,4 @@
-import SpaceInfo from 'web-pkg/src/components/sideBar/Spaces/SpaceInfo.vue'
+import SpaceInfo from 'web-pkg/src/components/SideBar/Spaces/SpaceInfo.vue'
 import { defaultPlugins, shallowMount } from 'web-test-helpers'
 
 const spaceMock = {
